+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 1.  [#3103](https://github.com/influxdata/chronograf/pull/3103): Add ability to clone dashboards
 1.  [#3080](https://github.com/influxdata/chronograf/pull/3080): Add tabular data visualization option with features
 1.  [#3120](https://github.com/influxdata/chronograf/pull/3120): Change :interval: to represent a raw influxql duration value
+1.  [#2832](https://github.com/influxdata/chronograf/pull/2832): Add paginated measurements API to server
 
 ### UI Improvements
 
@@ -22,6 +23,9 @@
 1.  [#3109](https://github.com/influxdata/chronograf/pull/3109): Display link to configure Kapacitor on Alerts Page if no configured kapacitor.
 1.  [#3111](https://github.com/influxdata/chronograf/pull/3111): Fix saving of new TICKscripts
 1.  [#3130](https://github.com/influxdata/chronograf/pull/3130): Fix appearance of cell Y-Axis titles
+1.  [#3129](https://github.com/influxdata/chronograf/pull/3129): Only add stateChangesOnly to new rules
+1.  [#3131](https://github.com/influxdata/chronograf/pull/3131): Fix 500s when deleting organizations
+1.  [#3137](https://github.com/influxdata/chronograf/pull/3137): Fixes issues with providing regexp in query
 
 ## v1.4.3.1 [2018-04-02]
 
@@ -31,7 +35,7 @@
 1.  [#3094](https://github.com/influxdata/chronograf/pull/3094): Save template variables on first edit
 1.  [#3101](https://github.com/influxdata/chronograf/pull/3101): Fix template variables not loading all values
 
-## v1.4.3.0 [2018-3-28]
+## v1.4.3.0 [2018-03-28]
 
 ### Features